add ability to calculate moving time

+++ PathFinder/Domain/GPSData.cs
@@ -9,9 +9,9 @@ namespace PathFinder.Domain
 
     public class WayPoint
     {
-        public Decimal Latitude { get; set; }
-        public Decimal Longitude { get; set; }
-        public Decimal? Elevation { get; set; }
+        public Double Latitude { get; set; }
+        public Double Longitude { get; set; }
+        public Double? Elevation { get; set; }
         public DateTime? TimeStamp { get; set; }
     }
 }
